rename: 공통으로 사용할 수 있을 것 같은 컴포넌트 파일 위치 변경 (in widgets)
