+++ src/widgets/CarouselComponent.tsx
