+++ src/widgets/IconComponent.tsx
